Rename location-emulation-icons to icons
Move icon components from packages/serve-sim/src/client/location-emulation-icons into packages/serve-sim/src/client/icons. No code changes were made to the files themselves (chrome.tsx, index.ts, modes.tsx, playback.tsx); this consolidates icon assets under a simpler directory name.
