+++ packages/serve-sim/src/client/icons/chrome.tsx
